refactor: move ChainConf to configuration
