+++ Cargo.toml
@@ -15,5 +15,5 @@ members = [
     "agents/processor",
     "tools/kms-cli",
     "tools/nomad-cli",
-    "tools/balance-exporter"
+    # "tools/balance-exporter"
 ]
