+++ chains/nomad-ethereum/src/macros.rs
@@ -123,12 +123,12 @@ macro_rules! boxed_trait {
     }};
     ($name:ident, $abi:ident, $trait:ident, $($n:ident:$t:ty),*)  => {
         #[doc = "Cast a contract locator to a live contract handle"]
-        pub async fn $name(conn: Connection, locator: &ContractLocator, signer: Option<Signers>, timelag: Option<u8>, $($n:$t),*) -> color_eyre::Result<Box<dyn $trait>> {
+        pub async fn $name(conn: nomad_xyz_configuration::chains::ethereum::Connection, locator: &ContractLocator, signer: Option<Signers>, timelag: Option<u8>, $($n:$t),*) -> color_eyre::Result<Box<dyn $trait>> {
             let b: Box<dyn $trait> = match conn {
-                Connection::Http { url } => {
+                nomad_xyz_configuration::chains::ethereum::Connection::Http { url } => {
                     boxed_trait!(@http url, timelag, $abi, signer, locator, $($n),*)
                 }
-                Connection::Ws { url } => {
+                nomad_xyz_configuration::chains::ethereum::Connection::Ws { url } => {
                     boxed_trait!(@ws url, timelag, $abi, signer, locator, $($n),*)
                 }
             };
