+++ configuration/src/chains/ethereum.rs
@@ -1,0 +1,25 @@
+//! Ethereum/EVM configuration types
+
+/// Ethereum connection configuration
+#[derive(Debug, serde::Deserialize, Clone, PartialEq)]
+#[serde(tag = "type", rename_all = "camelCase")]
+pub enum Connection {
+    /// HTTP connection details
+    Http {
+        /// Fully qualified string to connect to
+        url: String,
+    },
+    /// Websocket connection details
+    Ws {
+        /// Fully qualified string to connect to
+        url: String,
+    },
+}
+
+impl Default for Connection {
+    fn default() -> Self {
+        Self::Http {
+            url: Default::default(),
+        }
+    }
+}
