+++ nomad-base/src/settings/secrets.rs
@@ -38,9 +38,8 @@
 //!     }
 //! }
 
-use crate::ChainConf;
 use color_eyre::Report;
-use nomad_xyz_configuration::agent::SignerConf;
+use nomad_xyz_configuration::{agent::SignerConf, ChainConf};
 use serde::Deserialize;
 use std::collections::HashMap;
 use std::{fs::File, io::BufReader, path::Path};
